Initial commit
Have downloaded Matlab Gill code that I got from Adam. Modified to save
pdf plots (instead of relying on GUI, hard to run on Fletcher) and
verified that it runs.

Started "gillscratch" ipython notebook to translate the code into Python.
Got all the way to the first figure, but it doesn't look the same as the
Matlab version. Will have to print test output from both versions to
isolate the problem.

	new file:   gill.m
	new file:   gill01.pdf
	new file:   gill03.pdf
	new file:   gillscratch.ipynb