Fixed problems with
It turned out the Matlab version was producing different results from the
Python version for two main reasons:
(1) FFT for 2D matrix works along columns in Matlab but rows in Python.
(2) Initializing zero arrays resulted in complex numbers being cast
    to reals in Python.
Have fixed these errors, and now figures are the same. Also made
Figure 3 in the iPython notebook and verified this one works too.

Seems like Matlab on the Fletcher environment
can only export solid or dotted lines to PDF. Changed dashed
to dotted to verify that plots in the Python version are the same
as those in the Matlab version.

What to do next? Probably:
-Re-run Matlab code with different parameters to save the rest of the plots
-Do these plots in scratch script as well (or a 2nd script--gillscratch2?)
-Start converting into a Python module that can be imported
-Start running TRACMIP cases

	modified:   gill.m
	modified:   gill01.pdf
	new file:   gill01.png
	modified:   gill03.pdf
	new file:   gill03.png
	modified:   gillscratch.ipynb