More work on Gill model
In the "GillForTRACMIP.ipynb" notebook, made plots showing
geopotential height perturbation and wind quivers,
for two separate perturbations and the superposition
of the responses.

In gill.py, added an alternate function specifying a mass source
that is Gaussian in y but invariant in x over the continent.
Made plots using this version too.

	modified:   GillForTRACMIP.ipynb
	modified:   gill.py